requirements.txt
fastapi
uvicorn
sqlalchemy
transformers
gradio
tensorflow
tf-keras
python-multipart
aiofiles
numpy
Pillow
chromadb
torch